+++ src/memory/Memory.hpp
@@ -6,8 +6,8 @@
 class Memory
 {
 public:
-    uint8_t read(const uint8_t address) const;
-    void write(const uint8_t address, const uint8_t val);
+    uint8_t read(const uint16_t address) const;
+    void write(const uint16_t address, const uint8_t val);
     
 private:
     // 0x0000 - 0x01FF: Interpreter space
